+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.1.0 (2020-11-27)
+
+* Re-organization of variables
+
+### Fixes
+
+The cloudsql instance settings are now passed through the `settings` variable making it more in line with the terraform resource itself
+
 ## v1.0.0 (2020-11-23)
 
 * Initial release